refactor: move classes to their respective packages i.e ui,network, and models
