+++ app/src/androidTest/java/com/aboiyon/mycolletions/MainActivityInstrumentationTest.java
@@ -10,6 +10,8 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
 
+import com.aboiyon.mycolletions.ui.MainActivity;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
