+++ app/src/main/java/com/aboiyon/mycolletions/network/BookStoreApi.java
@@ -1,13 +1,14 @@
-package com.aboiyon.mycolletions;
+package com.aboiyon.mycolletions.network;
+
+import com.aboiyon.mycolletions.models.BookStoreSearchResponse;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
-import retrofit2.http.Query;
 
 public interface BookStoreApi {
     @GET ("search/{query}")
-    Call<BookStoreSearchResponse> searchBooks( @Path("query") String query);
+    Call<BookStoreSearchResponse> searchBooks(@Path("query") String query);
     @GET("new")
     Call<BookStoreSearchResponse> getNewBooks();
     @GET("books/{category}/{page}")
