+++ app/src/main/java/com/aboiyon/mycolletions/ui/MainActivity.java
@@ -1,17 +1,10 @@
-package com.aboiyon.mycolletions;
+package com.aboiyon.mycolletions.ui;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.aboiyon.mycolletions.databinding.ActivityMainBinding;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
     private ActivityMainBinding binding;
