+++ app/src/main/res/layout/activity_books.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.aboiyon.mycolletions.BooksActivity">
+    tools:context="com.aboiyon.mycolletions.ui.BooksActivity">
     <ProgressBar
         android:id="@+id/progressBar"
         android:layout_width="wrap_content"
