+++ app/src/test/java/com/aboiyon/mycolletions/MainActivityTest.java
@@ -4,6 +4,8 @@ import static org.junit.Assert.assertTrue;
 
 import android.widget.TextView;
 
+import com.aboiyon.mycolletions.ui.MainActivity;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
